refactor(fax): rename Hooks → SinchEvents and align event type names
