+++ src/Sinch/Fax/SinchEvents/IFaxSinchEvent.cs
@@ -1,0 +1,15 @@
+using System.Text.Json.Serialization;
+using Sinch.Core;
+
+namespace Sinch.Fax.SinchEvents
+{
+    [JsonInterfaceConverter(typeof(FaxSinchEventConverter))]
+    public interface IFaxSinchEvent
+    {
+        /// <summary>
+        ///     The type of the Fax Sinch event.
+        /// </summary>
+        [JsonPropertyName("event")]
+        public FaxEventType Event { get; }
+    }
+}
